Added new post screens
Added A "new post choose picture " screen, a "choose filter" screen, a "guided ALTogether screen" with a button to a "custom ALTogether screen, which both lead to a "caption" screen. I left the navigation to caption screens as a "next" button in upper right hand corner, but that can change.